refactor(parser): move main parser from runtime to jit
